refactor: move downloadState to src/util so the web can share it
